+++ src/main/java/paths/ShortestPathResult.java
@@ -3,10 +3,10 @@ package paths;
 import java.util.List;
 
 public class ShortestPathResult {
-    public float d;
+    public double d;
     public List<Integer> path;
 
-    public ShortestPathResult(float d, List<Integer> path) {
+    public ShortestPathResult(double d, List<Integer> path) {
         this.d = d;
         this.path = path;
     }
